requirements.txt
allure-pytest==2.9.43
allure-python-commons==2.9.43
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.4
colorama==0.4.4
idna==3.2
iniconfig==1.1.1
kafka-python==2.0.2
packaging==21.0
pluggy==1.0.0
py==1.10.0
pyDiameter==0.18.7.15
pyparsing==2.4.7
pytest==6.2.5
requests==2.26.0
six==1.16.0
toml==0.10.2
urllib3==1.26.6
